available_locales should return config from rails (if defined)

+++ test/test_helper.rb
@@ -20,13 +20,14 @@ end
 # Transactional fixtures do not work with Selenium tests, because Capybara
 # uses a separate server thread, which the transactions would be hidden
 # from. We hence use DatabaseCleaner to truncate our test database.
-DatabaseCleaner.strategy = :truncation
+DatabaseCleaner.strategy = :transaction
 
 DatabaseCleaner.clean_with :truncation
 
 class ActiveSupport::TestCase
+  self.use_transactional_fixtures = true
   setup do
-    Redis.new.flushall
+    Redis.new.flushall if Lit.key_value_engine == 'redis'
     Lit.init.cache.reset
   end
 end
@@ -39,13 +40,16 @@ class ActionDispatch::IntegrationTest
   self.use_transactional_fixtures = false
 
   setup do
+    DatabaseCleaner.strategy = :truncation
     I18n.backend.reload!
+    DatabaseCleaner.start
   end
 
   teardown do
     DatabaseCleaner.clean       # Truncate the database
     Capybara.reset_sessions!    # Forget the (simulated) browser state
     Capybara.use_default_driver # Revert Capybara.current_driver to Capybara.default_driver
+    DatabaseCleaner.strategy = :transaction
   end
 end
