+++ test/unit/lit/localization_test.rb
@@ -4,7 +4,7 @@ module Lit
   class LocalizationTest < ActiveSupport::TestCase
     test "does not create version upon creation" do
       I18n.locale = :en
-      Redis.new.flushall
+      Redis.new.flushall if Lit.key_value_engine == 'redis'
       assert_no_difference 'Lit::LocalizationVersion.count' do
         Lit.init.cache.reset
         assert_equal 'English translation', I18n.t('scope.text_with_translation_in_english')
@@ -13,7 +13,7 @@ module Lit
 
     test "does create new version upon update via model" do
       I18n.locale = :en
-      Redis.new.flushall
+      Redis.new.flushall if Lit.key_value_engine == 'redis'
       assert_difference 'Lit::LocalizationVersion.count' do
         Lit.init.cache.reset
         assert_equal 'English translation', I18n.t('scope.text_with_translation_in_english')
